Split ssa configuration out of ssa.cpp and put it into a separate header
